Remove set_vlan : only smrt.py script
